Move electron authentication code
